RA1/IL1.3/agente-avionica/requirements.txt
langchain
langchain-openai
langchain-community
chromadb
pypdf
python-dotenv
openai
